+++ modules/game_actions.rb
@@ -11,19 +11,4 @@ module GameActions
       print "------------" if i == 2 || i == 5
     end
   end
-  def game_start
-    # Welcome screen - call show_title method
-    show_title
-    # Draw initial board - initialize board
-    @game_board = Board.new
-    generate_board(@game_board.board)
-    
-    # Create players1 and 2 - initialize both players
-    player_one = Player.new("Player1","X")
-    player_two = Player.new("Player2","O")
-    
-    # Run a loop wherein players1 and 2 will each take turns until one wins or it"s a draw - game_cycle
-    game_cycle(player_one,player_two)
-  end
-
 end
